chatbot_backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
qdrant-client
psycopg2-binary
python-frontmatter
markdown
beautifulsoup4